raptrcontainer/requirements.txt
-e ..
Django==2.1.2
django-bootstrap3==11.0.0
django-filter==2.0.0
django-money==0.14.3
django-phone-field==1.6.0
django-tables2==2.0.2
phonenumbers==8.9.16
Pillow==5.3.0
py-moneyed==0.7.0
pytz==2018.6
setuptools==39.1.0
